backend/requirements.txt
fastapi[all]
uvicorn
sqlalchemy
passlib[bcrypt]
python-jose[cryptography]
python-multipart
python -m uvicorn app.main:app --reload
